# docs/swagger.yaml
basePath: /v1
definitions:
...
  models.DeleteStudent:
    properties:
      mssv:
        type: string
    required:
    - mssv
    type: object
...
  models.Scheduler:
    properties:
      amount:
        type: integer
      base:
        type: string
      capacity:
        type: integer
      classId:
        type: string
      className:
        type: string
      customerId:
        type: string
      doorId:
        description: DoorID          *string `gorm:"type:varchar(256);" json:"doorId"`
        type: integer
      employeeId:
        type: string
      endClassTime:
        type: integer
      endDate:
        type: string
      id:
        type: integer
      lecturerId:
        type: string
      lecturerName:
        type: string
      role:
        description: value in ["employee", "student", "customer"]
        type: string
      roomId:
        type: string
      roomName:
        type: string
      roomRow:
        type: string
      startClassTime:
        type: integer
      startDate:
        type: string
      status:
        type: string
      studentId:
        type: string
      userId:
        type: string
      weekDay:
        type: integer
    type: object
...
  models.Student:
    properties:
      email:
        type: string
      id:
        type: integer
      keypadPass:
        type: string
      major:
        type: string
      mssv:
        type: string
      name:
        type: string
      phone:
        type: string
      rfidPass:
        type: string
      schedulers:
        items:
          $ref: '#/definitions/models.Scheduler'
        type: array
    required:
    - mssv
    type: object
...
  models.SwagCreateStudent:
    properties:
      email:
        type: string
      keypadPass:
        type: string
      major:
        type: string
      mssv:
        type: string
      name:
        type: string
      phone:
        type: string
      rfidPass:
        type: string
    required:
    - mssv
    type: object
...
  models.UserSchedulerReq:
    properties:
      amount:
        type: integer
      base:
        type: string
      capacity:
        type: integer
      classId:
        type: string
      className:
        type: string
      customerId:
        type: string
      doorId:
        description: DoorID          *string `gorm:"type:varchar(256);" json:"doorId"`
        type: integer
      doorlockAddress:
        type: string
      employeeId:
        type: string
      endClassTime:
        type: integer
      endDate:
        type: string
      gatewayId:
        type: string
      id:
        type: integer
      lecturerId:
        type: string
      lecturerName:
        type: string
      role:
        description: value in ["employee", "student", "customer"]
        type: string
      roomId:
        type: string
      roomName:
        type: string
      roomRow:
        type: string
      startClassTime:
        type: integer
      startDate:
        type: string
      status:
        type: string
      studentId:
        type: string
      userId:
        type: string
      weekDay:
        type: integer
    required:
    - doorlockAddress
    - gatewayId
    type: object
  utils.ErrorResponse:
    properties:
      errorMsg:
        type: string
      msg:
        type: string
      statusCode:
        type: integer
    type: object
host: http://iot.hcmue.space:8002
info:
  contact:
    name: Mr. Khai
  description: This is API document for DMS backend server
  title: Backend API
  version: "1.0"
paths:
...
  /v1/student:
    delete:
      consumes:
      - application/json
      description: Delete student using "mssv" field. Send deleted info to MQTT broker
      parameters:
      - description: Student MSSV
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.DeleteStudent'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Delete Student By MSSV
    patch:
      consumes:
      - application/json
      description: Update student, must have correct "id" and "mssv" field. Send updated
        info to MQTT broker
      parameters:
      - description: Fields need to update a student
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Student'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Update Student By ID
    post:
      consumes:
      - application/json
      description: Create student
      parameters:
      - description: Fields need to create a student
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.SwagCreateStudent'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Student'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Create Student
  /v1/student/{msnv}/scheduler:
    post:
      consumes:
      - application/json
      description: Add scheduler that allows student open specific door. Send updated
        info to MQTT broker
      parameters:
      - description: Request with Scheduler, GatewayID, DoorlockAdress
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.UserSchedulerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Add Door Open Scheduler For Student
  /v1/student/{mssv}:
    get:
      description: find student info by student mssv
      parameters:
      - description: Student MSSV
        in: path
        name: mssv
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Student'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Find Student By MSSV
  /v1/students:
    get:
      description: find all students info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/models.Student'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Find All Student
swagger: "2.0"